chore(gb-9090): remove sdk 0.8, update to wasmtime 32
