+++ Cargo.toml
@@ -246,9 +246,9 @@ urlencoding = "2.1"
 uuid = "1.16.0"
 walkdir = "2"
 wasmparser = "0.228.0"
-wasmtime = "30"
-wasmtime-wasi = "30"
-wasmtime-wasi-http = "30"
+wasmtime = "32"
+wasmtime-wasi = "32"
+wasmtime-wasi-http = "32"
 webbrowser = "1"
 which = "7.0.2"
 winnow = "0.7"
