+++ crates/grafbase-sdk/wit/since_0_14_0/directive.wit
@@ -26,16 +26,10 @@ interface directive {
         items-range: tuple<u32, u32>
     }
 
-    record schema-directive {
-        subgraph-name: string,
-        name: string,
-        arguments: list<u8>
-    }
-
     record field-definition-directive {
         name: string,
         site: field-definition-directive-site,
-        // serialized in CBOR
+        // encoded in CBOR.
         arguments: list<u8>
     }
