--- crates/integration-tests/data/extensions/crates/resolver-08/.gitignore
@@ -1,3 +1,0 @@
-target
-build
-.build.lock
