--- crates/integration-tests/data/extensions/crates/resolver-08/extension.toml
@@ -1,8 +1,0 @@
-[extension]
-name = "resolver-08"
-version = "1.0.0"
-kind = "resolver"
-description = ""
-
-[permissions]
-network = true
