--- crates/wasi-component-loader/src/extension/api/since_0_8_0/instance/selection_set_resolver.rs
@@ -1,32 +1,0 @@
-use engine_error::GraphqlError;
-use futures::future::BoxFuture;
-use runtime::extension::Data;
-
-use crate::{
-    Error,
-    extension::{
-        SelectionSetResolverExtensionInstance,
-        api::wit::{ArgumentsId, Field, FieldId},
-    },
-};
-
-impl SelectionSetResolverExtensionInstance for super::ExtensionInstanceSince080 {
-    fn prepare(
-        &mut self,
-        _subgraph_name: &str,
-        _field_id: FieldId,
-        _fields: &[Field<'_>],
-    ) -> BoxFuture<'_, Result<Result<Vec<u8>, GraphqlError>, Error>> {
-        Box::pin(async { unreachable!("Not supported by this SDK") })
-    }
-
-    fn resolve_query_or_mutation_field(
-        &mut self,
-        _headers: http::HeaderMap,
-        _subgraph_name: &str,
-        _prepared: &[u8],
-        _arguments: &[(ArgumentsId, &[u8])],
-    ) -> BoxFuture<'_, Result<Result<Data, GraphqlError>, Error>> {
-        Box::pin(async { unreachable!("Not supported by this SDK") })
-    }
-}
