--- crates/wasi-component-loader/src/extension/api/since_0_8_0/mod.rs
@@ -1,3 +1,0 @@
-pub mod instance;
-pub mod types_impl;
-pub mod wit;
