+++ crates/wasi-component-loader/src/extension/api/since_0_9_0/wit/token.rs
@@ -1,3 +1,12 @@
 pub use super::grafbase::sdk::token::*;
 
 impl Host for crate::WasiState {}
+
+impl From<Token> for runtime::extension::Token {
+    fn from(value: Token) -> Self {
+        match value {
+            Token::Anonymous => Self::Anonymous,
+            Token::Bytes(items) => Self::Bytes(items),
+        }
+    }
+}
